cpufreq: interactive: Skip timer when policy->min == policy->max
When the minfreq is set to the maxfreq, there is nothing to really do: the
CPU is already running at its maxfreq. In this scenario, just do nothing
and re-arm the timer.

This scenario occurs when the screen is turned on due to the cpu_input_boost
driver (it sets policy->min to policy->max for a few seconds).

Change-Id: I0fd31eeca7311f1147bd505b9a9d779551db6156
Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Rohit Vashisht <rohitv88cf@gmail.com>