power: msm-core: Compile out temperature polling
The temperature polling used here is expensive; according to top, it
regularly eats up 10% of the CPU. Frequently spending 10% of CPU time on an
*energy efficiency* feature is nonsense, so we'll just use the default
temperature for power calculations so that we can still make use of
power-aware scheduling.

Signed-off-by: Sultanxda <sultanxda@gmail.com>

**[@dev-harsh1998]** : adapt for msm8916

Adapted for msm8952 as well with very mild changes at the end.

Signed-off-by: Rohit Vashisht <rohitv88cf@gmail.com>